+++ src/migrations/Install.php
@@ -7,6 +7,7 @@
 
 namespace craft\commerce\stripe\migrations;
 
+use craft\commerce\stripe\db\Table;
 use craft\db\Migration;
 use craft\helpers\MigrationHelper;
 
@@ -23,7 +24,8 @@ class Install extends Migration
      */
     public function safeUp(): bool
     {
-        $this->createTable('{{%stripe_customers}}', [
+        $this->archiveTableIfExists(Table::CUSTOMERS);
+        $this->createTable(Table::CUSTOMERS, [
             'id' => $this->primaryKey(),
             'userId' => $this->integer()->notNull(),
             'gatewayId' => $this->integer()->notNull(),
@@ -34,7 +36,8 @@ class Install extends Migration
             'uid' => $this->uid(),
         ]);
 
-        $this->createTable('{{%stripe_invoices}}', [
+        $this->archiveTableIfExists(Table::INVOICES);
+        $this->createTable(Table::INVOICES, [
             'id' => $this->primaryKey(),
             'reference' => $this->string(),
             'subscriptionId' => $this->integer()->notNull(),
@@ -44,7 +47,8 @@ class Install extends Migration
             'uid' => $this->uid(),
         ]);
 
-        $this->createTable('{{%stripe_paymentintents}}', [
+        $this->archiveTableIfExists(Table::PAYMENTINTENTS);
+        $this->createTable(Table::PAYMENTINTENTS, [
             'id' => $this->primaryKey(),
             'reference' => $this->string(),
             'gatewayId' => $this->integer()->notNull(),
@@ -56,19 +60,19 @@ class Install extends Migration
             'uid' => $this->uid(),
         ]);
 
-        $this->addForeignKey(null, '{{%stripe_customers}}', 'gatewayId', '{{%commerce_gateways}}', 'id', 'CASCADE');
-        $this->addForeignKey(null, '{{%stripe_customers}}', 'userId', '{{%users}}', 'id', 'CASCADE');
-        $this->addForeignKey(null, '{{%stripe_invoices}}', 'subscriptionId', '{{%commerce_subscriptions}}', 'id', 'CASCADE');
-        $this->addForeignKey(null, '{{%stripe_paymentintents}}', 'gatewayId', '{{%commerce_gateways}}', 'id', 'CASCADE');
-        $this->addForeignKey(null, '{{%stripe_paymentintents}}', 'customerId', '{{%stripe_customers}}', 'id', 'CASCADE');
+        $this->addForeignKey(null, Table::CUSTOMERS, 'gatewayId', '{{%commerce_gateways}}', 'id', 'CASCADE');
+        $this->addForeignKey(null, Table::CUSTOMERS, 'userId', '{{%users}}', 'id', 'CASCADE');
+        $this->addForeignKey(null, Table::INVOICES, 'subscriptionId', '{{%commerce_subscriptions}}', 'id', 'CASCADE');
+        $this->addForeignKey(null, Table::PAYMENTINTENTS, 'gatewayId', '{{%commerce_gateways}}', 'id', 'CASCADE');
+        $this->addForeignKey(null, Table::PAYMENTINTENTS, 'customerId', Table::CUSTOMERS, 'id', 'CASCADE');
 
-        $this->createIndex(null, '{{%stripe_customers}}', 'gatewayId', false);
-        $this->createIndex(null, '{{%stripe_customers}}', 'userId', false);
-        $this->createIndex(null, '{{%stripe_customers}}', 'reference', true);
-        $this->createIndex(null, '{{%stripe_invoices}}', 'subscriptionId', false);
-        $this->createIndex(null, '{{%stripe_invoices}}', 'reference', true);
-        $this->createIndex(null, '{{%stripe_paymentintents}}', 'reference', true);
-        $this->createIndex(null, '{{%stripe_paymentintents}}', ['gatewayId', 'customerId', 'transactionHash'], true);
+        $this->createIndex(null, Table::CUSTOMERS, 'gatewayId', false);
+        $this->createIndex(null, Table::CUSTOMERS, 'userId', false);
+        $this->createIndex(null, Table::CUSTOMERS, 'reference', true);
+        $this->createIndex(null, Table::INVOICES, 'subscriptionId', false);
+        $this->createIndex(null, Table::INVOICES, 'reference', true);
+        $this->createIndex(null, Table::PAYMENTINTENTS, 'reference', true);
+        $this->createIndex(null, Table::PAYMENTINTENTS, ['gatewayId', 'customerId', 'transactionHash'], true);
 
         return true;
     }
@@ -78,12 +82,12 @@ class Install extends Migration
      */
     public function safeDown(): bool
     {
-        MigrationHelper::dropAllForeignKeysOnTable('{{%stripe_invoices}}', $this);
-        MigrationHelper::dropAllForeignKeysOnTable('{{%stripe_customers}}', $this);
-        MigrationHelper::dropAllForeignKeysOnTable('{{%stripe_paymentintents}}', $this);
-        $this->dropTable('{{%stripe_customers}}');
-        $this->dropTable('{{%stripe_invoices}}');
-        $this->dropTable('{{%stripe_paymentintents}}');
+        MigrationHelper::dropAllForeignKeysOnTable(Table::INVOICES, $this);
+        MigrationHelper::dropAllForeignKeysOnTable(Table::CUSTOMERS, $this);
+        MigrationHelper::dropAllForeignKeysOnTable(Table::PAYMENTINTENTS, $this);
+        $this->dropTable(Table::CUSTOMERS);
+        $this->dropTable(Table::INVOICES);
+        $this->dropTable(Table::PAYMENTINTENTS);
 
         return true;
     }