+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Stripe API version are now set on all client requests. ([#340](https://github.com/craftcms/commerce-stripe/issues/340))
 - Added support for fonts in the Stripe Elements. ([#258](https://github.com/craftcms/commerce-stripe/issues/258))
+- Added `craft\commerce\stripe\db\Table`.
+- Stripe for Commerce is more defensive about DB tables already existing during install.
 
 ## 5.0.7 - 2025-03-19